pyproject.toml
[tool.poetry]
name = "teachers-service"
version = "0.1.0"
description = ""
authors = ["Damego <damego.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
motor = "^3.5.1"
pymongo = "^4.8.0"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
